// src/types/rs-types/rs-soknadstype.ts
export enum RSSoknadstype {
    SELVSTENDIGE_OG_FRILANSERE = 'SELVSTENDIGE_OG_FRILANSERE',
    OPPHOLD_UTLAND = 'OPPHOLD_UTLAND',
    ARBEIDSTAKERE = 'ARBEIDSTAKERE',
    REISETILSKUDD = 'REISETILSKUDD',
    ARBEIDSLEDIG = 'ARBEIDSLEDIG',
    BEHANDLINGSDAGER = 'BEHANDLINGSDAGER',
    ANNET_ARBEIDSFORHOLD = 'ANNET_ARBEIDSFORHOLD'
}

export type RSSoknadstypeType =
    'SELVSTENDIGE_OG_FRILANSERE' |
    'OPPHOLD_UTLAND' |
    'ARBEIDSTAKERE' |
    'REISETILSKUDD' |
    'ARBEIDSLEDIG' |
    'BEHANDLINGSDAGER' |
    'ANNET_ARBEIDSFORHOLD'